(Not related to DSL) - fix many tangles caused by graph being a sub-package of support.management  - make `graph` a top-level package
