+++ spring-integration-core/src/main/java/org/springframework/integration/graph/EndpointNode.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.support.management.graph;
+package org.springframework.integration.graph;
 
 /**
  * Base class for all endpoints.
