+++ spring-integration-core/src/main/java/org/springframework/integration/graph/LinkNode.java
@@ -14,17 +14,43 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.support.management.graph;
+package org.springframework.integration.graph;
 
 /**
- * Nodes implementing this interface are capable of emitting errors.
+ * Represents a link between nodes.
  *
  * @author Gary Russell
  * @since 4.3
  *
  */
-public interface ErrorCapableNode {
+public class LinkNode {
 
-	String getErrors();
+	private final int from;
+
+	private final int to;
+
+	private final Type type;
+
+	public LinkNode(int from, int to, Type type) {
+		this.from = from;
+		this.to = to;
+		this.type = type;
+	}
+
+	public int getFrom() {
+		return this.from;
+	}
+
+	public int getTo() {
+		return this.to;
+	}
+
+	public Type getType() {
+		return this.type;
+	}
+
+	public enum Type {
+		input, output, error, discard, route
+	}
 
 }
