+++ spring-integration-core/src/main/java/org/springframework/integration/graph/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides classes related to the runtime object graph.
+ */
+package org.springframework.integration.graph;
