--- spring-integration-core/src/main/java/org/springframework/integration/support/management/graph/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Provides classes related to the runtime object graph.
- */
-package org.springframework.integration.support.management.graph;
